Signals implementation in charts.component.ts, open-positions.component.ts  and live-quotes.component.ts.  ChangeDetectionStrategy OnPush mode
