+++ src/app/pages/dashboard/open-positions/open-positions.component.ts
@@ -1,34 +1,17 @@
-import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { OpenPosition } from 'app-core/models/open-position.model';
-import { Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
 import { OpenPositionFacade } from 'app-core/facades/open-position.facade';
+import { toSignal } from '@angular/core/rxjs-interop';
 
 @Component({
   selector: 'app-open-positions',
   templateUrl: './open-positions.component.html',
   styleUrls: ['./open-positions.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class OpenPositionsComponent implements OnInit, OnDestroy {
-  openPositions: OpenPosition[] = [];
-  private destroy$ = new Subject();
-  constructor(
-    private openPositionsFacade: OpenPositionFacade,
-    private cd: ChangeDetectorRef
-  ) {}
-
-  ngOnInit() {
-
-    this.openPositionsFacade
-      .getAllOpenPositions()
-      .pipe(takeUntil(this.destroy$))
-      .subscribe((value) => {
-        this.openPositions = value;
-        this.cd.detectChanges();
-      });
-  }
-  ngOnDestroy() {
-    this.destroy$.next(true);
-    this.destroy$.complete();
-  }
+export class OpenPositionsComponent  {
+  private openPositionsFacade = inject( OpenPositionFacade)
+  positions = toSignal<OpenPosition[]>(
+    this.openPositionsFacade.getAllOpenPositions()
+  );
 }
