--- src/app/pages/dashboard/rates-charts/rates-charts.component.html
@@ -1,4 +1,0 @@
-<ng-container *ngIf="options">
-  <div echarts [options]="options" class="demo-chart"></div>
-</ng-container>
-
